✅ Fix all TypeScript build errors - Build now passes
- Fix request.ip property issue in orders/route.ts
- Add proper type assertion for balance validation
- Fix type safety in data redaction functions
- All TypeScript errors resolved

Build Status: SUCCESSFUL ✅
Only warnings remain (non-blocking)

+++ lib/blockchain/balanceValidation.ts
@@ -67,7 +67,7 @@ export async function validateWalletBalance(
       abi: USDC_ABI,
       functionName: 'balanceOf',
       args: [walletAddress as `0x${string}`]
-    })
+    }) as bigint
     
     // USDC has 6 decimals
     const balanceInUSDC = parseFloat(formatUnits(balance, 6))
